// App/TS-REDUX/src/model/APIResponsesTypes.ts
import store from "./store";

export type Animal = "bird" | "cat" | "dog" | "rabbit" | "reptile";

export interface IPet {
	id: number;
	name: string;
	animal: Animal;
	description: string;
	breed: string;
	images: string[];
	city: string;
	state: string;
}

export interface IPetAPIResponse {
	numberOfResults: number;
	startIndex: number;
	endIndex: number;
	hasNext: boolean;
	pets: IPet[];
}

export interface IBreedAPIResponse {
	animal: Animal;
	breeds: string[];
}

export type RootType = ReturnType<typeof store.getState>;